requirements.txt
matplotlib
torch
Pillow
torchvision
numpy
scikit-learn
pandas
python3-tk
